requirements.txt
wakeonlan~=3.0.0
python-dotenv~=1.0.0
pythonping~=1.1.4